components/section.test.tsx
import React from 'react'
import { render, screen } from '@testing-library/react'; // Import the screen object
import userEvent from '@testing-library/user-event'; // Import the userEvent module
import SectionComponent from './section'

test('SectionComponent renders correctly', () => {
  render(<SectionComponent type={''} appCodeData={undefined} onChange={function (data: any, dataKey: any): void {
    throw new Error('Function not implemented.')
  } } />)
  // Add your assertions here
})

test('SectionComponent renders title correctly', () => {
  const title = 'Test Title'
  render(<SectionComponent title={title} type={''} appCodeData={undefined} onChange={function (data: any, dataKey: any): void {
    throw new Error('Function not implemented.')
  }} />)
  // Add your assertions here
})

test('SectionComponent renders children correctly', () => {
  const children = <div>Test Children</div>
  render(<SectionComponent type={''} appCodeData={undefined} onChange={function (data: any, dataKey: any): void {
    throw new Error('Function not implemented.')
  }}>{children}</SectionComponent>)
  // Add your assertions here
})

test('SectionComponent collapses and expands on click', () => {
  render(<SectionComponent title="Test Title" type={''} appCodeData={undefined} onChange={function (data: any, dataKey: any): void {
    throw new Error('Function not implemented.')
  }} />)
  const legend = screen.getByText('Test Title')
  userEvent.click(legend)
  // Add your assertions here
})